refactor(*): move time-step-data to config and remove config function
