+++ pollination/honeybee_vtk/config.py
@@ -3,19 +3,34 @@ from pollination_dsl.function import Function, command, Inputs, Outputs
 
 
 @dataclass
-class Config(Function):
-    """Write a config file to be consumed by other commands of honeybee-vtk."""
+class TimeStepData(Function):
+    """Write a time step data JSON file to be consumed by the TimeStepImages."""
 
-    input_file = Inputs.file(
-        description='Path to the input json file.',
-        path='input.json'
+    time_step_file = Inputs.file(
+        description='A file generated by an annual recipe that has all the time steps.'
+        ' an example of such a file is sun-up-hours.txt generated by the'
+        ' direct-sun-hours recipe.',
+        path='time_step_file.txt'
+    )
+
+    periods_file = Inputs.file(
+        description='A file that contains the periods and color associated with each'
+        ' period.',
+        path='peridos.json'
+    )
+
+    name = Inputs.str(
+        description='Name of the time step data file to be written.',
+        default='time_step_data',
     )
 
     @command
-    def write_config(self):
-        return 'honeybee-vtk config input.json --name config --folder-path .'
+    def time_step_data(self):
+        return 'honeybee-vtk config time-step-data time-step-data'\
+            ' --time-step-file time_step_file.txt --periods-file periods.json'\
+            ' --file-name {{self.name}} --folder target_folder'
 
-    config = Outputs.file(
-        description='Config.json file.',
-        path='config.json'
+    output_file = Outputs.file(
+        description='Created file.',
+        path='target_folder/{{self.name}}.json'
     )
