+++ tests/config_test.py
@@ -1,8 +1,8 @@
-from pollination.honeybee_vtk.config import Config
+from pollination.honeybee_vtk.config import TimeStepData
 from queenbee.plugin.function import Function
 
 
-def test_write_config():
-    function = Config().queenbee
-    assert function.name == 'config'
+def test_time_step_data():
+    function = TimeStepData().queenbee
+    assert function.name == 'time-step-data'
     assert isinstance(function, Function)
